fix dk_cfsHtmlTpl field definition
there was no possibility to select a custom mod_caroufredsel_ticker.html5 template, as the ce_caroufredsel instead of mod_caroufredsel_ticker template would be used. The problem arose with a change in contao which does not provide an empty option in the select boxes anymore. before, with an empty select the custom template from /templates would automatically be used, as the default is defined in ModuleCarouFredSelTicker.php:35